Fixed Tomcat Regex
Test case of fixed regex is here: https://regex101.com/r/TfdeSA/1/ The original regex did not work in Burp, and also did not work in the regex tester.

Are there other test cases this regex should have worked for? If so, this change will break those. If not, it's possible a number of these are broken due to unescaped forward slashes.