Read whole test file at once
In the previous version the file was read line by line and all lines were
concatenated, which resulted in the removal of end-of-line characters. Since
all newlines in the test file are removed, this does not make a realistic test
when the matching text contains a newline. Read the whole file including
newlines instead.