Detect Telerik version comments
Telerik provides a .NET web framework. Sometimes they put the exact version in a HTML comment. It doesn't really specify that it's the Telerik version, so I have put the certainty on "Firm" instead of "Certain".

In the testResponse.txt, the `# Telerik` is a comment and not part of the server response.